[Bifrost] Remove read/read_opt from loglet contract
This removes read/read_opt from loglet contract and provides equivalent methods in LogletWrapper that are used only in tests. This greatly simplifies the requirement for building a loglet, removes duplicate code, and makes it easy to evolve the read API.

Moreover, loglet tests now use LogletWrapper and LogletWrapper's readstream to (a) make use of the test-only read/read_opt API, and (b) include LogletWrapper in the scope of testing.

Test Plan: Unit tests