[Bifrost] BackgroundAppender implementation
Introduces a new way to enqueue messages to Bifrost. This PR adds `bifrost.create_background_append(...).start()` that starts a background appender. The background appender automatically batches appends and provides flexible control over draining it or auto-destruction when all senders are dropped (see `detach()` on `AppenderHandle`).

With this, `ActionEffectHandler` makes use of this new ability and will not block the partition processor loop when writing effects to bifrost.