Allow state to be deleted from non-existent services
This will allow the `state clear` CLI command to work even for services
that don't yet exist. We change the state modification API to allow
services that don't exist when the request is to remove state (and not
insert anything else). This means that it would be a no-op for truly
non-existent services, and a useful op for ex-services.