Add Cleaner task to clean completed invocations (#1845)
* A bunch of changes regarding time.

Expanded the data structure StatusTimestamps that tracks changes within the InvocationStatus. Because those timestamps are currently not agreed between replicas, the functions are marked unsafe and should really be used only in cases where those timestamps don't influence the deterministic business logic of the PP. Also only the NeoInvocationStatus stores those fields, the old invocation status will never fill those.

Now CompletedInvocation propagates the `CompletionRetention`. This field is filled with a `Duration::MAX` for the old invocation status, causing the addition to overflow (thus not causing the cleanup in the new cleaner task).

"Time is what the clock says" A. Einstein

* Implement the new Cleaner task. Also stop scheduling the cleanup timer for NeoInvocationStatus.