Resizing PartitionId and LogId and types
This PR reduces the size of PartitionId to a conservative u16 and LogId to u32. This attempts to fix a category of problems we often see when working with those generously-sized types. We would like to have tighter control on how many one of the types would fit into the other, for instance, we can share a bifrost log-id space with multiple clusters. Additionally, with a small partition-id space, we always have the option to expand it later, but with the current u64, it'd be very difficult to go the other way around.

This PR comes with a small risk of incompatibility if anyone has was able to run with more than 65k partitions on a single node. I don't think that's feasible with the amount of resources that it'd require, so, I deem this transition as (almost) safe. We would be performing further compatibility checks prior to release, but according to my local tests, I've not observed any issues.