Address proc-macro-error advisory RUSTSEC-2024-0370
The only dependency that currently needs `proc-macro-error` is `err-derive` used by `arrow_convert` which has not been updated yet. For our use, we only rely on it for restate's CLI so the risk is minimal but we will upgrade when a fix is available upstream. For the time being, we'll ignore this advisory` is `err-derive` used by `arrow_convert` which has not been updated yet. For our use, we only rely on it for restate's CLI so the risk is minimal but we will upgrade when a fix is available upstream. For the time being, we'll ignore this advisory.