[LogServer] Initial implementation for store path
Note 1: This PR contains many todos an various incomplete bits that will be worked out in later PRs. This is highly experimental code and will be refined in the next iterations but I wanted to avoid accumulating code locally.

Note 2: This enables log-server by default, so it's imperative that this isn't included in v1.1 release.

When looking at this, consider the overall design strategy and don't get hang up on the nits or smaller details as they're likely to change anyway.

What this achieves:
- Starts log-server by default (role is now enabled by default)
- Introduces Store/Release network messages for the log-server
- Implements pipelined Store messages from sequencers, no backfill ability.
- LogState management
- Provisional trim infrastructure