Disable log-server role by default to fix v1.0.x compatibility
This change disables the log-server role by default to avoid breaking. Additionally, if log-server role is enabled, log-server will not start unless the binary is built with `-F replicated-loglet` to protect the public release from starting the server. This change is necessary to maintain compatibility with v1.0.x.

This change also attempts to detect the presence of log-store and asks the user to remove it manually. We don't remove it automatically to avoid accidental data wipes if accidental rollbacks took place from future versions back to this.