Initial ReplicationChecker implementation
This introduces ReplicationChecker utility that performs write and f-majority quorum checks. Note that the current implementation is not LocationScope aware and will only consider node-level scope.

The utility is designed to allow attaching any attribute type to nodes in the nodeset and run the check against the set of nodes that match a predicate against those attributes.