Change the IngressSubmitNotification implementation.
Before we were sending the attached invocation id and the submitted notification id, and based on whether those two are different or not, the ingress would reply to /send with status code accepted or previously accepted. We now just notify back the ingress with `is_new_invocation: bool`. This also includes some changes in the tests, which now don't need to create two different ids anymore.

Replace all the usages of the old InvocationId/InvocationUUID methods with the new methods. No semantic changes here.