Deterministic Invocation ID generation (#1996)
* Introduce the new deterministic ID generation.

`InvocationUUID::generate` is the relevant code, everything else is removing/shuffling around methods we don't need anymore.
Add a couple of tests for the deterministic id generation in InvocationId.

* Datafusion sys_inbox.created_at was the only SQL field that previously used the TimestampAwareId, and this field was queried by the CLI (although not really used?).

This is a breaking change of the SQL interface.

* Change the IngressSubmitNotification implementation.

Before we were sending the attached invocation id and the submitted notification id, and based on whether those two are different or not, the ingress would reply to /send with status code accepted or previously accepted. We now just notify back the ingress with `is_new_invocation: bool`. This also includes some changes in the tests, which now don't need to create two different ids anymore.

Replace all the usages of the old InvocationId/InvocationUUID methods with the new methods. No semantic changes here.

* Fix bug with hashing of IdempotencyId (this affected only the ingress get output/attach functionality when using virtual objects)

* Use &str instead of `ByteString`

* Feedback