[Bifrost] Unify log-server request headers and global_tail exchange in log-server responses
This introduces `LogServerRequestHeader` and ensures we exchange `known_global_tail` on log-server responses. Additionally, we removed the dedicated tracking map for the global tail and it's now part of the `LogletState`. This simplifies the code and paves the way for a future optimization where we periodically persist the last known global tail on log-servers for more efficient tail repair in edge cases.