[Bifrost] GetDigest message in replicated loglets
Digest as a summary of offset-range presence on a certain node. A a new request message `GetDigest` is introduced to the log-server to allow an efficient scanning of a range of offsets to check which offsets reside on this server. This process will be used as part of the tail repair to determine (a) whether a record meets the replication requirements and (b) where to fetch a certain record if we need to re-replicate.