Networking is no longer a dependency for MetadataManager
MetadataManager no long needs `Networking` as dependency. It'll keep track of the WeakConnection of where we observed the updated version, then it'll attempt to use this connection to fetch the update (no need to re-establish the connection with the peer, or block the metadata manager loop). If fetching from this connection fails, metadata manager falls back to syncing from metadata store as usual.

This also sets a timeout while waiting for metadata store fetches to allow the system to recover by letting other nodes share their metadata before the next tick comes in.