Enhance the `restatectl describe log` command (#2067)
This change renames the previous `restatectl dump cluster-state` command to `restatectl partitions status`. It now correctly prints out the chain segments for each log id specified on the command line, with an option to also display the loglet parameters for replicated logs. We support displaying individual or ranges of log ids, as well as various options to limit the number of segments to display per chain.