Facilities to communicate per-role status across the system (#2104)
This introduces a facility to emit and merge status updates for every role. At the moment status protos are simple enums and I expect those to change to a more elaborate mergeable types (hence support for the `Merge` trait).

This also adds a few new fields to `GetIdent`. It now returns:
- age in seconds (since daemon started)
- metadata versions
- per-role status
- node-status is now reflective of the main NodeSvc network interface status