Introduces NetworkServerBuilder to enable inversion of grpc service registration
This allows roles/services to add services to node-rpc-server on startup with no requirement for Node to know about service construction requirements.

Part of this PR temporarily removes QueryStorage API from NodeSvc as it should not belong here. This should either move to its own gRPC service (thanks to the new registration mechanic) or to be removed completely if this is to be replaced by node-to-node networking.