LogServer observability gRPC service stub
This introduces as new gRPC service stub for logservers, the service interface is only a stub at the moment, future PRs will include the implementation and the concrete message formats.

LogServerSvc is registered as expected. This doesn't have access to the request pump at the moment. I expect us to create a channel between the grpc handler and RequestPumb to request access to LogletWorkers and such.
```console
2024-10-24T15:57:07.926744Z DEBUG restate_core::network::server_builder
  Registering gRPC service to node-rpc-server
    svc: "restate.log_server.LogServerSvc"
```