Introduce TaskKind::RpcResponse which runs on the default runtime
In order to make sure that rpc responses originating from the PartitionProcessor
are being sent even if the PartitionProcessor runtime is being shut down, this
commit changes the PartitionProcessor to use the TaskKind::RpcResponse for sending
rpc responses back. The TaskKind::RpcResponse runs on the default runtime.

An alternative could have been to await that all spawned tasks complete before
shutting down the partition processor runtime. But there aws no easy way to do this
for non-local tasks. Also it creates the risk that we prevent the shut down from
happening if one of those tasks does not react to the shut down signal.

This fixes #2270.