Implement Attach/Get output in service-protocol (#2266)
* WIP remove this and replace with git subtree update

* Add parsing of AttachInvocationEntryMessage and GetInvocationOutputEntryMessage

* Add new field to AttachInvocation to reply immediately when the invocation is in-flight (essentially implementing the get invocation output semantics we already have for the ingress), handle it and test it.
This commit also contains a "special case" for get invocation output, that translates the CompletionResult with NOT_READY_ERROR into Empty. This was done to have a good API for the service-protocol, but for the time being I don't want to change the `InvocationResponse` command in the WAL protocol. We might reconsider this anyway when we'll get to the immutable log.

* Add OutboxMessage::AttachInvocation to propose Command::AttachInvocation in another PP

* Implement the new entry messages