Support comma-separated list of roles in restate-server
Summary:

Allows restate-server to start with `--roles=admin,worker` in addition to the existing method of repeating the option `--roles=admin --roles=worker`. Note that `--role` is also a supported alias.

This also fixes the Display impl of `Role` to print the kebab-case representation of the `Role` type for a consistency.

Test Plan:

Starting restate-server with a mixture of style works as expected:
```
cargo run --bin restate-server -- --roles=log-server,admin --role=metadata-store
...
...
2024-11-19T11:24:35.488926Z INFO restate_node
  My Node ID is N0:3
    roles: admin | metadata-store | log-server
    address: http://127.0.0.1:5122/
```