[TaskCenter] Stage 1 of task-center clean-up
TaskCenter has been an incredibly useful tool but it has grown organically and is spiraling out of control. This PR is the first step to break it down into smaller modules in preparation for a more ergonomic API for its users.

Changes in this PR are mechanical and does cause any behavioural changes.