Fix problem with run future not being cancellation safe
This commit fixes the problem where the run future of the LeaderState was
losing action effects if handle_action_effects would be awaiting and another
tokio::select arm would trigger in the PP event loop. The solution is to
pass the action effects back to the PP event loop and then handle them in
the select arm.