[TaskCenter][5/n] Introduces restate_core::test macro
Introduces a new macro that creates a default TaskCenter in tests. This should replace usage of `tokio::test` (the code is mostly based on it) and it provides access to `TaskCenter::current()` within the test code.

This also introduces a transitional type `TestCoreEnv2`, a clone of `TestCoreEnv` but without TaskCenter. After migration, this will be renamed to `TestCoreEnv` again.