[TaskCenter] Requests cancellation of managed runtimes
This lets task center request the graceful cancellation of the root-future of managed runtimes. Note that it'll not wait for the shutdown to complete. If this is needed, then partition-processor managed will need to wait for the result to come when it receives the cancellation signal on its own cancellation token.