[Provisional] Enabled WAL compression
This also adds compression for log-server's L0/L1 levels (WAL is STD as it's the only supported type, and L0/1 is LZ4 [provisional choice])
This should be considered a provisional change to give us some time to test this in different environments and validate its performance characteristics. I've observed roughly 50% steady-state disk usage reduction since our WAL is fairly compressable. The impact on CPU depends on the hardware/os combination but so far I'm positive that we'll move ahead with this.

Let's leave this baking for some time, and I'll continue experimenting with different compression types for SST files. I'll make this a configuration option if we decided it's something we want to keep.