Let MetadataStore use the node grpc server
In order to make the MetadataStore use the node grpc server, the grpc
server needs to be started first. Since a few grpc services depend on
a node having joined the cluster (in order to obtain a NodeId), this
commit disables these until they can be used.

To disable Tonic services, this commit introduces the TonicServiceFilter
which takes a Tonic service and a predicate. The predicate allows to
check on a status to decide whether to forward the request or fail the
request with tonic::Status.

This fixes #2411.