[Networking] Enable and tune h2 keep-alive on the server-end of connections
h2 keep-alive is very important to detect frozen streams, we only had this enabled on client-side of connections so we didn't always detect broken network or frozen processes. This also tunes the default such that we can detect frozen processes in 5-10s. The biggest impact comes from bifrost's remote sequencers being able to detect that their in-flight appends can/should be retried, otherwise the'll be stuck indefinitely even if the loglet has been sealed already (albeit this happens on a higher level).
- Changes the max frame size to 32MB (default is 4MB) to allow for larger batches
- Fixes a small lineage issue in tracing individual connection tasks (more to follow on the larger issue in subsequent PRs)
- Changes how shutdown is designed for connections. Connections now wait for 5 seconds for graceful sub-streams drain before dropping the connection
- Minor renames and logging