[Bifrost] Introduce ErrorRecoveryStrategy control
This PR introduces the new parameter to appender's API to let API users control the behaviour when appends to the current loglet face persistent failures. There is a new feature (`auto-extend` in restate-bifrost which controls the default behaviour).

No logic changes happen in this PR, in future PRs, bifrost will respect this input and will configure its behaviour accordingly.