[sql] Extract a partition key from the filters
This commit adds a simple partition pruning, based on the passed
filters. With this commit, we are able to prune partitions based on either
an invocation_id or a service_key (these columns named differently
in different tables) and also a partition_key.
In addition to pruning the partitions, we also set the scan to be
more precise based on the partition_key.