Use cargo hakari to automate workspace-hack
This PR integrates workspace-hack (via [cargo hakari](https://crates.io/crates/cargo-hakari)), this fixes the huge space amplification that we had with rocksdb builds, and it building/checking individual crates becomes significantly faster after the initial build.

The cost of this:
- Slightly longer initial build time (but I didn't really notice much difference on my machine)
- I removed io-uring optional feature (easy to add back when we decide to give it another try) from rocksdb
- After adding new third-party dependencies, make sure you run `cargo hakari generate`, don't worry, CI will catch if you didn't.

This is a non-destructive change, and running with `cargo hakari` disable should still work if this is needed.