Replace the usage of GenerationalNodeId with PlainNodeId and StorageId
Using the generational node id as the OmniPaxos NodeId is a bit problematic because
it will get bumped on every restart. If we fail while joining an existing OmniPaxos
cluster and before persisting the configuration, we will try joining with a different
id. If we allow higher generations to be accepted, then this is problematic because we
can't distinguish between this situation and the loss of the disk. If we are strict
about the id equality, then we might get the system stuck because the restarted node
cannot join the OmniPaxos cluster which might expect its previous generation to be
present. Therefore, this commit introduces the StorageId which is a random number
stored in the RocksDbStorage. Whenever we try to join an OmniPaxos cluster we send
this StorageId along with the PlainNodeId. Based on the StorageId we can detect if
a node lost its disk because it would generate a new one.