[Locality] Location-aware NodeSetChecker
This introduces a new sophisticated and performance-optimized nodeset checker for bifrost flexible quorums. It uses Arc+RwLock (parking-lot) to make this compatible with multi-threaded tokio runtimes. I initially designed it with Rc<RefCell<..>> and an `unsafe impl Send` but after thinking long and hard about it, there was no easy way to ensure that the safety guarantees won't break with future refactoring. Parking lot's RwLock has excellent performance in uncontended case which is _always_ going to be the case here.

The new nodeset checker has extensive test coverage