Send known leader information back to GrpcMetadataStoreClient
If a metadata store cannot answer a request because it is not the leader but it
knows who the leader is, then it sends this information back to the
GrpcMetadataStoreClient as part of the Status Metadata map. The client will use
this information to pick the next metadata store to query on the next attempt.