Harden ClusterController by not panicking on errors
Instead of escalating errors from the ClusterController service which
ultimately leads to a panic, this commit catches those errors and logs
them on warning in the hope that a retry will fix the problem. The down-
side is that a buggy cluster controller will never step down so that
another cluster controller can take over as the leader.

This fixes #2276.