Add LogId, database name, and column family name to Partitioon
Summary:
Extend the Partition to inlcude more partition related attributes.
This now include:
- Log Id
- Database Name
- Column Family Name

So far these values are defaults and there is no way to change them
but this will probably change in the future

This PR only adds those fields to the partition structure but the rest
of the system still not using it. This should happen in a
follow up PR.