Fix updating NodesConfiguration after Metadata store reconfigurations
Setting nodes that are not part of the configuration to Standby is problematic
because it might set some nodes that are trying to join the cluster to Standby.
If this happens, then the nodes will stop trying to join the cluster. Hence,
we are only setting those nodes to Member which are part of the current configuration.
Note, currently this does not change the NodesConfiguration since we use the Member
state to tell nodes to join a cluster in the first place.