Remove inner retry mechanism from GrpcMetadataServerClient
Instead of letting the client retry, it is now the responsibility of the
user to retry if she wants. This gives more flexibility and avoids the negative
effet of prolonged retry delays if there is a nesting of retry policies. Only
if the response let's us learn about a known leader, we will retry immediately.