[Bifrost][Auto-Recovery] Appender can reconfigure loglet if sealed
In certain scenarios, if the appender is known to be running with "permission" to extend the chain, it'll run out of patience and will attempt to perform reconfiguration itself. This works really well, but it might compete with a cluster controller that thinks a different placement should take place. The auto-recovery mechanism only kicks in after 3 seconds (not configurable, but will be in a later PR) and we can see that the appender will attempt to reconfigure the loglet if it's sealed.

At the moment, this will get the system out of trouble in certain scenarios.

```
// intentionally empty
```