Allow mark_the_node_as_writeable to tolerate retry attempts
In the presence of network errores, mark_the_node_as_writeable must be able to
tolerate retry attempts. This commit changes the behvaiour of this method to
only fail on the first attempt if the StorageState is not Provisioning. Moreover,
it adds a basic retry policy for retryable errors.

This fixes #2563.