Make restatectl cluster provision and cluster config set more flexible
This commit makes the cluster provision and cluster config set command more
flexible. It now allows to specify only a partial set of values and it will
take the remaining values from the current cluster configuration/default
values.

Additionally, this commit introduces the new type NodeSetSize to enforce the
upper limit of 128 across the configuration and API calls.

This fixes #2604 and #2561.