Disable debug_assertions in release builds
Debug assertions are not meant to be used in release builds, most crates use `cfg(debug_assertions)` as a proxy to detect whether they are built in debug mode or not, this include `tracing` among many others. The side effect of having debug_assertions enabled is (a) performance penalty and (b) TRACE level being enabled in release build because `tracing` can't respect the release max logging filter feature flag if it doesn't know we are building in release mode.

```
// intentionally empty
```