Add a random value to the object store based metadata
This commit adds a random value into the serialized envelope
written into the object store, this is to guard against a
theoretical situation where both participants try to CAS
exactly the same value at the same time, hence producing the
same ETag (hash of the content), both might succeed.
To break the tie this commit adds a random u64 to the serialized
payload.