[Bifrost] RecordCache record swap efficiency
We were swapping records in the record cache with serialized ones, that hurts the use-case where the sequencer and the reader are co-located. The reason this was happening was because after the sequencer adds the "Typed" variant of the entry into cache, the co-located log-server would overwrite it.

This lets us gain 10-20% more throughput (YMMV).

```
// intentionally empty
```