Auto migrate PlainNodeId 0 to the forced node id or the next available PlainNodeId
This commit adds a migration path for the LocalMetadataServer which updates a 0 PlainNodeId to a
non-zero value in the NodesConfiguration. If the node that runs the local metadata server had the
node id 0, then the migration path will respect a potentially forced node id if it is set and not
yet taken. Otherwise, the migration path will pick the next available PlainNodeId.

This fixes #2550.