Allow to change the default provider to replicated
This commit allows users to change the default log provider to replicated even if the
cluster has been fully provisioned. This allows users to migrate from a single node
deployment with a local loglet to a replicated loglet and then a multi node deployment.
Additionally, this commit enables the LogsController to react to default provider changes.
If the new default provider is the replicated loglet, then the logs controller will reconfigure
all logs.