[LogServer] Support for data-loss transition
- Log-server auto-transitions into `data-loss` if it detected that its marker is lost
- log-server rejects to start if the not has lost its marker and not in provisioning state

Note that this still doesn't mark the node as data-loss if corruption was detected in rocksdb. Current behaviour is that the log-server will fail to start but the node remains in read-write.

If the node detected marker loss, this is what it prints on the log
```
on rs:worker-8
2025-02-14T09:40:24.606443Z WARN restate_log_server::service
  Detected data loss for log-server of my node N1. The log-server marker is missing, storage-state was read-write, will transition into `data-loss`
on rs:worker-11
  in restate_log_server::service::provision_node
2025-02-14T09:40:24.609743Z ERROR restate_core::task_center
  Shutting down: task 4 failed with: Node cannot start a log-server on N1, it has detected that it has lost its data. storage-state is `data-loss`
    kind: SystemBoot
    name: "init"
on rs:worker-5
2025-02-14T09:40:24.609794Z WARN restate_core::task_center
  ** Shutdown requested
    reason: task init failed and requested a shutdown
```

And on restart:
```
2025-02-14T09:42:54.172461Z ERROR restate_core::task_center
  Shutting down: task 4 failed with: A previous generation of my node N1 has marked this log-server's storage-state as `data-loss`.If this is a new node, please choose a different node-name to provision it as a new node
    kind: SystemBoot
    name: "init"
```

The node will start if log-server role was removed, albeit the cluster will spew many networking warnings due to log-server messages not being registered. Handling this scenario is a separate issue.