[restatectl] Support for setting/viewing log-server storage-state
- adds `replicated-loglet servers` or (list-servers)
- adds `replicated-loglet set-storage-state --node-id=4 --storage-state=read-only`

Note that this is restrictive at the moment, except that it liberally allows transition from `read-write` into `read-only` even if this leads the cluster in a state where no further nodesets with enough writeable nodes to generate nodesets. A check for this will in a future follow-up
Screenshots with the behaviour in comments.

```
// intentionally empty
```